+++ teaching/cs455/changelog.md
@@ -4,4 +4,7 @@
 - SQL Lab #1: Creation of passenger database; simple queries
 - SQL Lab #2: Tougher, aggregation queries
 - SQL Lab #3: show them node.js + sqlite. Connect it to previous two labs somehow.
-- Revamp project grading scheme. Make them run scrum.
+- Project to be converted to being more open-ended
+
+# Change Log
+- Normalization homework converted to programming